I added this to the wrong branch on accident.
I have not changed this code since before the deadline (October 26). I just accidentally added a file to this repository instead of the 'hw3' branch I meant to add to, but I am deleting it now. I will also archive this branch so I don't do this again. You can verify that the hw is unchanged via my commit history.